Matched files
Matched `StarPieceMother`
Matched `GameDataHolder`
Matched `StarPieceAlmsStorage`

+++ include/Game/System/GameDataHolder.hpp
@@ -28,7 +28,7 @@ public:
     bool isOnGameEventValueForBit(const char*, int) const;
     void setGameEventValueForBit(const char*, int, bool);
     s32 getPictureBookChapterCanRead() const;
-    u16 getPictureBookChapterAlreadyRead() const;
+    u32 getPictureBookChapterAlreadyRead() const;
     void setPictureBookChapterAlreadyRead(int);
     void setRaceBestTime(const char*, u32);
     u32 getRaceBestTime(const char*) const;
@@ -54,14 +54,14 @@ public:
     void addStockedStarPiece(int);
     s32 setupSpinDriverPathStorage(const char*, int, int, int, f32*);
     void updateSpinDriverPathStorage(const char*, int, int, f32);
-    s32 getStarPieceNumGivingToTicoSeed(int) const;
+    u16 getStarPieceNumGivingToTicoSeed(int) const;
     u32 getStarPieceNumMaxGivingToTicoSeed(int) const;
     void addStarPieceGivingToTicoSeed(int, int);
     bool isCompleteMarioAndLuigi() const;
     bool isPassedStoryEvent(const char*) const;
     void followStoryEventByName(const char*);
     void resetAllData();
-    s32 makeFileBinary(u8*, u32);
+    u32 makeFileBinary(u8*, u32);
     bool loadFromFileBinary(const char*, const u8*, u32);
 
     /* 0x00 */ GameEventFlagChecker* mEventFlagChecker;
